Fix missing expung, remember learned ham, added movehamto
expung was missing after learnspambox.
Mails passed threw learnhambox is added to the pastuidsfile.
Added an option for moving learned ham to a different folder.

User case for movehamto:

INBOX
INBOX.spam
INBOX.spamtraning
INBOX.notspam

Periodicaly run:
./isbg.py --imaphost imap.excample.com --noreport
./isbg.py --imaphost imap.excample.com --learnspambox INBOX.spamtraning --learnhambox INBOX.notspam --movehamto INBOX --learnthendestroy --expunge

User can move mails that are incorrectly moved to spam to the notspam box, once it has been learned it is moved back to his inbox where is belongs.
User can move spam that isn't detected to the spamtraning where it will be learned, and then deleted.
The user can either delete spam in the spam folder or move it to the spamtraning folder.