Significant speed up in fetching UIDs, Remove ham from pasteuids
Use of imap search to get the uids of messages less then the thresholdsize is 100x faster (725 mails took 3.75 it now takes 0.0374 sec) then the previously used code.
UIDs are unique only to the folder so adding in ham before it was moved would cause incorrect UIDs to be added.