Quick fix for spamc error exit, remove lock file on spamc error, exit if spamd is missconfigured
Spamc returns "0/0\n" not "0/0", this should be fixed in a way that it will work if spamc is later changed to return just "0/0" or maybe "0/0\r\n" "0/0\r" on a different system.
Use errorexit when exiting on Spamc error.
Exit when traning if spamd is not setup properly, and give a hit on how to fix it.