Set stderr loghandler in `isbg_run`
Setting the log handler that logs everything to stderr only in the
`isbg_run` method meant for when isbg is run as command / top level
script.

When isbg is run as a library, the caller is responsible for setting up
log output. If we add a log handler this can lead to messages being
duplicated.