avoid that spamassassin is executed twice for one spam mail
In spamproc.py each mail was first processed by test_mail() function. test_mail() is returning if the mail is SPAM or no SPAM.
If the mail has been classified as SPAM feed_mail() is executed to attach the SPAM report to the mail. But feed_mail() will execute spamassassin again for the same email.

Due to the high ressource usage of spamassassin that's not very efficient. Now spamassassin will only called once.